% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyzeCDMelt}
\alias{analyzeCDMelt}
\title{Calculation of Fraction Unfolded from Native to Denature State}
\usage{
analyzeCDMelt(data, folded_temp, unfolded_temp, pch = 16, ...)
}
\arguments{
\item{data}{list outputted from \link{plotCDMelt} function.}

\item{folded_temp}{temperature, in celsius, where the biomolecule is assumed fully-folded.}

\item{unfolded_temp}{temperature, in celsius, where the biomolecule is assumed fully-unfolded.}

\item{pch}{the shape of the data points.}

\item{...}{passing arguments to \link{plot} function.}
}
\description{
This function calculates the fraction unfolded, equilibrium constants, and natural log of
equilibrium constants from the native state to the denatured state. Afterwards, a plot of the fraction unfolded as
a function of temperature is given.
}
\examples{
\dontrun{

protein <- importCD()

protein220nm <- plotCDMelt(protein, wavelength=220)

analyzeCDMelt(protein220nm, folded_temp=15, unfolded_temp=95)


}



}
