% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCD.R
\name{importCD}
\alias{importCD}
\title{Import Files From Jasco J-810 Circular Dichroism (CD) Spectropolarimeter}
\usage{
importCD(files)
}
\arguments{
\item{files}{path to where csv files  are located. If missing, the
user is prompted to select a folder interactively.}
}
\value{
a list containing a data frame and an atomic vector
}
\description{
This is an importing function that takes the CSV files from the Jasco J-810
machine and organizes all the CD Melt spectra data at each temperature into a large data frame.
}
\examples{

\dontrun{

#selecting file interactively
protein <- importCD()


}


}
