% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoCD.R
\name{infoCD}
\alias{infoCD}
\title{Extract Information from a CD Melting Data File}
\usage{
infoCD(file, params = FALSE)
}
\arguments{
\item{file}{a single CSV file from the Jasco J-810 CD experiment}

\item{params}{prints information about the experiment}
}
\value{
an atomic vector
}
\description{
The info function extracts experimental parameters in the CSV files exported
from Jasco J-810 Circular Dichroism Spectropolarimeter
}
\examples{

\dontrun{
 details <- info()

}

}
