% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDMelt.R
\name{plotCDMelt}
\alias{plotCDMelt}
\title{Melting Data Plot}
\usage{
plotCDMelt(data, wavelength, title, col = "red", pch = 16, ...)
}
\arguments{
\item{data}{a class object 'cd'}

\item{wavelength}{specific wavelength of interest for plotting multiplicities as a function
of temperature}

\item{col}{color of the points (col="red" is default)}

\item{pch}{shape of the points (pch=16 is default)}

\item{...}{passing arguments to \link{plot} function}
}
\description{
A plot of the ellipticity values at a given wavelength over
temperature.
}
\seealso{

}
