% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDSpectra.R
\name{plotCDSpectra}
\alias{plotCDSpectra}
\title{Overlay of All CD Spectra}
\usage{
plotCDSpectra(data, ...)
}
\arguments{
\item{data}{an object of class 'cd'}

\item{...}{passing arguments to \code{plot} function.}
}
\value{
a plot
}
\description{
overlay all circular dichroism spectras
}
\examples{

\dontrun {

#import data
protein <- importCD()

#create spectras
plotCDSpectra(protein)

}



}
