% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDVantHoff.R
\name{plotCDVantHoff}
\alias{plotCDVantHoff}
\title{Calculating Thermodynamic Parameters using Van't Hoff Equation}
\usage{
plotCDVantHoff(
  data,
  folded_temp,
  unfolded_temp,
  digits = 3,
  remove.temp = NULL,
  title = "Van't Hoff Equation",
  ...
)
}
\arguments{
\item{data}{an object of class 'vh'}

\item{folded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-folded.
Default folded_temp is your first temperature reading.}

\item{unfolded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-unfolded.
Default unfolded_temp is your final temperature reading.}

\item{digits}{how many digits shown for your thermodynamic parameters.}

\item{remove.temp}{remove specified data points from the linear regression model.}

\item{...}{passing arguments to \link{plot} function.}
}
\description{
Assuming a two-state model of melting a protein, this function automates the calculation of fraction unfolded
from a fully-folded state to fully-unfolded state. At each temperature, equilibrium constants are derived and plotted based off of the
Van't Hoff equation. From this linear model, enthalpy, entropy, Gibbs free energy of the system, and melting temperature of the
system are estimated.
}
\examples{
\dontrun {



protein <- importCD() \%>\%
plotCDMelt(wavelength=210) \%>\%
plotCDVantHoff(folded_temp=10, unfolded_temp=95)



}

}
