% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamics.R
\name{thermodynamics}
\alias{thermodynamics}
\title{Calculating Thermodynamic Parameters using van't Hoff (VH) Equation}
\usage{
thermodynamics(data, start, end, digits = 3, legend = TRUE, ...)
}
\arguments{
\item{data}{an object of class 'vh'; outputted from \link{analyzeCDMelt} function.}

\item{start}{temperature, in celsius, of the start of the transition curve.}

\item{end}{temperature, in celsius, of the end of the transition curve.}

\item{digits}{how many digits shown for thermodynamic parameters.}

\item{legend}{shows the thermodynamic parameters using the Van't Hoff. Set legend=FALSE to
hide the legend.}

\item{...}{passing arguments to \link{plot} function.}
}
\description{
This function requires the user to specify the start and end temperature (in Celsius)
range for the transition curve, where a plot of natural log of equilibrium constants versus 1/Temperature (in Kelvin).
}
\examples{
\dontrun{
protein <- importCD() \%>\%
plotCDMelt(pwavelength=210) \%>\%
analyzeCDMelt()
}


}
