% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamicsCD.R
\name{thermodynamicsCD}
\alias{thermodynamicsCD}
\title{Calculating Thermodynamic Parameters}
\usage{
thermodynamicsCD(
  data,
  fully_folded_temp,
  fully_unfolded_temp,
  remove.temp = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{data}{a list containing ellipticities, temperatures in Celsius, and information about the experiment.
(i.e, output of plotCDMelt() function)}

\item{fully_folded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-folded.}

\item{fully_unfolded_temp}{temperature (in Celsius) at which the biomolecule is assumed denatured.}

\item{remove.temp}{remove data points from the Van't Hoff plot (see example below)}

\item{digits}{how many digits shown for slope, y-intercept, enthalpy and entropy.}

\item{...}{passing the arguments to \link{plot} function}
}
\description{
This function automates the calculation of fraction unfolded from the ellipicities in a two-state model,
equilibrium constant, and thermodynamics of the system. Additionally, a Van't Hoff plot is provided with enthalpy,
entropy, and melting temperature values.
}
