% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCD.R
\name{importCD}
\alias{importCD}
\title{Import Circular Dichroism Files From Jasco J-810}
\usage{
importCD(files)
}
\arguments{
\item{files}{path to where csv files for the CD data are located. If missing, the
user is prompted to select files interactively.}
}
\value{
a list containing a data frame and an atomic vector
}
\description{
This is a user-friendly, importing function that takes CSV files exported from
the Jasco J-810 Circular Dichroism Spectropolarimeter.
}
\examples{

\dontrun{

#selecting file interactively
protein <- importCD()

}


}
