% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDMelt.R
\name{plotCDMelt}
\alias{plotCDMelt}
\title{Thermal Melting Plot of Ellipticities}
\usage{
plotCDMelt(data, wavelength, col = "red", pch = 16, ...)
}
\arguments{
\item{data}{the object outputted from \link{importCD} function.}

\item{wavelength}{specific wavelength of interest for plotting ellipticities as a function
of temperature.}

\item{col}{color of the points on the graph.}

\item{pch}{shape of the points on the graph.}

\item{...}{passing arguments to \link{plot} function.}
}
\description{
a plot of the ellipticity values (at user-specified wavelength), as a function of
temperature.
}
\examples{
\dontrun {

protein <- importCD()
protein220nm <- plotCDMelt(protein, wavelength=220)

}

}
