% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDMelt.R
\name{plotCDMelt}
\alias{plotCDMelt}
\title{Melting Data Plot}
\usage{
plotCDMelt(data, wavelength, col = "red", pch = 16, ...)
}
\arguments{
\item{data}{a class object 'cd'}

\item{wavelength}{specific wavelength of interest for plotting ellipticities as a function
of temperature.}

\item{col}{color of the points on the graph.}

\item{pch}{shape of the points on the graph.}

\item{...}{passing arguments to \link{plot} function.}
}
\description{
a plot of the ellipticity values at a given wavelength over
temperature.
}
\examples{
\dontrun {

protein <- importCD() \%>\% plotCDMelt(wavelength=220)


}

}
