% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDVantHoff.R
\name{plotCDVantHoff}
\alias{plotCDVantHoff}
\title{Calculating Thermodynamic Parameters using Van't Hoff Equation}
\usage{
plotCDVantHoff(
  data,
  folded_temp,
  unfolded_temp,
  digits = 3,
  remove.temp = NULL,
  title,
  ...
)
}
\arguments{
\item{data}{an object of class 'vh'}

\item{folded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-folded}

\item{unfolded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-unfolded}

\item{digits}{how many digits shown for slope, y-intercept, enthalpy and entropy}

\item{remove.temp}{remove data points from the Van't Hoff plot}

\item{...}{passing arguments to \link{plot} function}
}
\description{
This function automates the calculation of fraction unfolded from the ellipicities in a two-state model,
equilibrium constant, and thermodynamics of the system. Additionally, a Van't Hoff plot is provided with enthalpy,
entropy, and melting temperature values.
}
