% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VH_old.R, R/plotCDVH.R
\name{plotCDVH}
\alias{plotCDVH}
\title{Calculating Thermodynamic Parameters using Van't Hoff Equation}
\usage{
plotCDVH(
  data,
  start,
  end,
  digits = 3,
  main = "van't Hoff Plot",
  legend = TRUE,
  ...
)

plotCDVH(
  data,
  start,
  end,
  digits = 3,
  main = "van't Hoff Plot",
  legend = TRUE,
  ...
)
}
\arguments{
\item{data}{an object of class 'vh'; outputted from \link{analyzeCDMelt} function.}

\item{start}{temperature, in celsius, of the start of the transition curve.}

\item{end}{temperature, in celsius, of the end of the transition curve.}

\item{digits}{how many digits shown for thermodynamic parameters.}

\item{...}{passing arguments to \link{plot} function.}

\item{folded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-folded.
Default folded_temp is your first temperature reading.}

\item{unfolded_temp}{temperature (in Celsius) at which the biomolecule is assumed fully-unfolded.
Default unfolded_temp is your final temperature reading.}

\item{remove.temp}{remove specified data points from the linear regression model.}
}
\description{
Assuming a two-state model of melting a protein, this function automates the calculation of fraction unfolded
from a fully-folded state to fully-unfolded state. At each temperature, equilibrium constants are derived and plotted based off of the
Van't Hoff equation. From this linear model, enthalpy, entropy, Gibbs free energy of the system, and melting temperature of the
system are estimated.

This function requires the user to specify the start and end temperature (in Celsius)
range for the transition curve, where a plot of natural log of equilibrium constants versus 1/Temperature (in Kelvin).
}
\examples{
\dontrun {



protein <- importCD() \%>\%
plotCDMelt(wavelength=210) \%>\%
plotCDVantHoff(folded_temp=10, unfolded_temp=95)



}

\dontrun {

protein <- importCD() \%>\%
plotCDMelt(wavelength=210) \%>\%
analyzeCDMelt()

}

}
