% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{spectra}
\alias{spectra}
\title{Melting Temperature Graph of All CD Spectras}
\usage{
spectra(data, show.legend = TRUE, ...)
}
\arguments{
\item{data}{object outputted from \link{infoCD} function.}

\item{show.legend}{shows the legend of the melting temperature. Set show.legend=FALSE to hide.}

\item{...}{passing arguments to \code{plot} function.}
}
\value{
a color-coded plot of each CD spectras on one graph.
}
\description{
This function returns overlays all the CD  in one graph.
}
\examples{

\dontrun{
#import data
protein <- importCD()
#create spectras
plotCDSpectra(protein)
}
}
